requirements.txt
pandas>=2.0.0
pyarrow>=14.0.0
requests>=2.31.0
pydantic>=2.0.0
python-binance>=1.0.19
pyyaml>=6.0
